Major progress toward panning.
You can now pan around an infinite canvas and draw where you
like. When you save, only the current viewport is currently
captured. Likewise: loading is way broken (the bitmap isn't
even inserted in the draw stack, let alone load/save DrawOps
which is what really needs to be done anyway).

But it is a start.