Some early scaling work.
Invalidations are getting screwed up, as is rescaling after
translating. Strokes are not being unscaled appropriately so
it's tough to do detail work.

But it's a start.

The right solution is to replace all the
viewportScale/viewportTranslate variables with a prev and
current matrix. This will help when transforming the spots
and the invalidations.