Assume stylus is calibrated.
This means removing it from the entire pressure calibration
system, not only not being influenced by current calibration
settings but not feeding input to the calibration either.
The PressureCooker is now purely a TOOL_TYPE_FINGER
calibration system.

Also fixed a bug where the tool type on PressureCooker
output was not being set correctly.

Thanks to my niece Isabel Makulski for requiring I fix this. :)