Fixing too-many-undo's in tiles.
Here's something astonishing: off-by-one-errors are such a
common programming error that occasionally you make so many
of them that they cancel one another out and the behavior is
essentially correct. Only when you go back to fix or add
something that depends on the intended behavior do you
notice the problem, and then it becomes something of a
pest-control problem to fix all the errors.