Explicitly adapt to system decor.
android:fitsSystemWindows is a nice idea, but it totally
breaks when you have multiple views that each need to adapt
to the status and navigation bars (since the first View that
returns true for fitSystemWindows() aborts the traversal).

For many apps this isn't a problem because you can simply
contain all your views in a single layout that fits the
system windows, but since Markers has a Slate view that
extends underneath the decor we need to apply the fitting
somewhere else. The next logical approach would be to create
a single large layout to hold all the palettes and other UI,
but that introduces a lot of unnecessary overdraw.

Here, we nominate a single view (I chose LogoView) to catch
fitSystemWindows() and apply those insets to many views
onscreen that might need them. As it happens adding inset
padding to these views doesn't muck around with the overall
layout so this works pretty well in practice. (You might see
a similar technique going on in the AOSP sources for
Launcher3.)

While this fixes outstanding bugs with older versions of
Android, I think this technique (entering immersive mode
only when the palettes are gone) isn't all that useful in
practice. Rather, I think Markers just needs to run in
immersive mode all the time, or perhaps with a (gasp)
setting.